On application startup check for initilization and show setup page
When application starts different things will be checked. The frontend application
will show an initialization status page. When all checks are passed as positive
(migrations, admin user) a continue button will be presented which redirects the
user to the initial setup page.

After doing setup (which creates a config) the frontend is configured and will
show the normal frontend dashboard (with status information about database and
display).