Fix display status detection and on/off switch handling
The complete output of command needs to be compared, so we first need to use the
correct output line ("display=...") and include the ending newline character.

Separate handling for on and off switch: first check if switch off is reached
and skip checking for on switch - otherwise it will switched on and off again
and again.