Use Connection header to avoid keep-alive connections in ApiClient
Usually it is a good idea to use a keep alive connection because this will
speed up things. But for this few requests we have no real performance
problems.

Maybe this will also work around with "host not found" exceptions or connections
which are closed on the remote end but block the connection pool.