Improve wifi listing when in Access Point mode
Disable Access Point when starting up and it is already activated to be able
to run a success full wifi scan with available networks.

When activating the Access Point do a wifi scan and keep the available networks
in the network status. While Access Point is active only provide the cached
network list to be able to connect to them (no further scanning is possible
in this situation).