Override item_ptr to disable cascade deletion
When an item is deleted Django tries to delete all references including the
data from RankedItem model. But this is not a real model (managed=False), but
Django still tries to delete it.

Setting abstract=True will also not work because such a model can not be
serialized by the Django REST framework.

Using proxy=True is also not work because we have additional fields in this
model declared.

The only solution is to override the reference to the Item model explicitly
and disable delete cascading.