Keep track of running jobs and generate dynamic ID in Scheduler.trigger()
Using event listeners to keep track of running jobs in a list. When using trigger()
method to start a job generate a dynamic job ID with current ID and timestamp
appended.

The method running() can be used to check if a given job (by ID) is currently running
by the scheduler executor. The optional flag starts_with can be used to check for
jobs starting with given name (useful for job started with the trigger() method).