Make item update visible in frontend to show if currently running (button disabled)
Trigger the next update (and also the status submit job) by using the scheduler
to be able to check if job is running or not. The status is now provided in the
field "display.refresh" which is true in case the update is currently running.

Make the status of update job visible by disabling the refresh button in the
frontend.