Put current item into own file instead of using numeric indexed files in frontend
Separate the list of latest items from the one currently shown on the display. This
way it is possible to generate a new image including additional things (e.g.
message or icon hints, or just switch back to an older image currenlty available on
the device).

Also reduce the reload of the CSV file for feh to 2 seconds to make it refresh
"immediately" when changing.