Show latest item n frontend when starting up and items are available
When starting up frontend application and some items from previous processing
are still available just activate and show the latest item. Next update should
be scheduled on the next update interval.

Added new startup() method to visualization renderer which checks and starts
up the rendering application if required. This is used in the FrontendDevice,
which is now also keeping an instance of the visualization renderer.