Implement jobs for config (currently only one source update job)
This job is executed each minutes and checks if some source needs to
be updated. If one source is currently updated for a frame no other
sources of ths frame will be updated and next source from another
frame will be processed.

All sources will be processed in the order of their last update date,
descending.