Fix Process.exec_run() in combination with sudo check
When running in sudo mode the given command is checked if it can be used with current
user. If not, skip it. Otherwise run it.

This check is using the given command and adds a "-l" parameter to check the
availablility. This had two problems:
1. the original argument list is modified by adding the "-l" parameter but
should be untouched to execute the original command later
2. the response of the check command is returned instead of just none

So no sudo command was executed correctly.