Completely refactor startup initialization
Introduced a new phase for config setup (entered some settings) to be able
to decide if we have a dummy config or already a user config available.

Changed BaseSetupView and BaseFrontendView to have their own _process()
methods (name them unique and invoke them both). Add handling of
unfinished setup (config is there, but no values set up) to redirect
user to the setup pages.

Also update the check in the SetupView to check for completed setup
instead of config availability (can be unconfigured).