Introduce special messages for startup frontend page and use it when saving configuration
Using the "message" parameter can indicate which message should be shown to
the user.

Currently we use this when changing/saving the configuration and show the
user a custom message. When custom messages are used, the success message
will just get a "completed!" appended handle this in a generic way.