Completely refactory job / scheduler handling
Merge Jobs class into Scheduler class which is now used directly. This class
needs to be extended and implement configure() method to setup jobs.

Since method names change all the scheduler usages need to be updates to the
new interface.

Added a new run_job() method which directly run a custom job, which is not
registered in the scheduler. In contrast to add_job() which starts a
registered job.

All registered jobs can be enabled or disabled using the enalbe_jobs() or
disable_jobs() methods.