Fix OR conditions in frame/source update query of jobs
First, change OR to AND condition for global interval because it should only be used when
frame is configured to use global internval.

Second, change resulting query to restrict global filters with AND (e.g. frame disabled
or display disabled) and combine the interval queries with OR (only one of them have
to match).