Refactor rendering and activation process in frontend (separate file renderer from others)
The FilesystemFrontendRenderer is now the main renderer which stores the items into
files (images, json, etc.) and keeps meta data available. This is is now handled by
the process() method. Other renderers can implement this method, but mainly there
is currently no need for it.

Further, the activate() method will get the item generated by the FilesystemFrontendRenderer
to activate the item. This item contains all required information to be able to
activate it (e.g. for visualization copy the item to the current item to show them
up).