Refactor finishing process (rendering and frontend item)
The FilesystemFrontendRenderer is now not responsible to put the images into the
file system. Instead it is only responsible to activate a given item by copying
the item to the current file.

Generating the items and storing them into the file system is now moved to the
rendering process of FrontendDevice directly. It is now responsible to cache
the items in the file system.

The FilesystemFrontendRenderer is now only responsible to activate a given item
by copying the item to the current file.

All renderers have to implement the activate() method to activate a given
item. They all will get now a FrontendItem object which contains all the
information about an item (the item itself, the image data and meta data,
the preview data and meta data).