Change utils.Filesystem.file_match() to return files and dirs and optionall recursive
It can be specified if only file name should be returned or only directory names
by using the new flags (files, dirs). Additionally a recurse option can be used
to recursively search directories.

Still the pattern need only to match the file names (maybe will be changed in
future to also match the dir names when returning them).

The prefix parameter can be used to prefix all files/dirs returned. When not
specified the files/dirs are relative to the given path.