Directly read thumbail file to create image instead of temporary non-existant file in display item API
When POSTing a display item update the thumbnail transmitted will be put into a
file object, which is at this time not persisted in the file system. But it is
a File object using a BytesIO buffer, which can be read.

Just read the files data and create an image from it to get the meta information.