Handle existing and new files in Data model correctly
Use name as identifier if file of Data model is new or already exists. In case it
is new just use read() method, otherwise use path of file object to read content.

The update() method will now also not close the file. The file is closed then the
model is saved.