Adjust Filesystem.file_rotate: keep only count items, add start, add reverse
This method should only keep given amount of file count. It will no also support
a starting point (defaults to 0 - so complete file list will be rotated). It
will also support a reverse mode to move the files from larger numers to lower
instead from lower to larger numbers (which is the default)