Introduce tree-based structures / models based on treebeard for plugins
Added new BaseManager / BaseQuerySet for normal models and TreeManager / TreeQuerySet
for tree-based structures. BaseManager is used in the BaseModel class for all models,
and TreeManager for tree-based models (introduced a TreePluginModel which can be
used for plugin models).

A new serializer class was also introduced for plugin models to reflect the plugin
structure when exporting / importing. Tree-based structures can use the
TreePluginImplementationSerializer.