Add additional body to card_tree_list_item() template for additional content (e.g. finishing delete)
This parameter can be used to put additional content into the item directly instead
of adding content to the children part (e.g. by just putting content after the
template call).

The deletion dialog have to be directly in item content and not in the children
part to make it visible (children are collapsed) and useable when clicking on
delete button.