Add "first-child" action for tree based views and use in config finishing view
Move sortable initialization to own template which can be used - also for nested
lists. This template have to be used for the root list and will automatically be
added if sortable flag is set in card_tree_list_item() template.

Sortable is now also restricted to the grip handle to avoid moving elements around
just by scrolling down in mobile view.