Introduce new "dir" source plugin to read files from media directory
New configuration setting MEDIA_PATH is the root path for the "dir" source plugin and
defaults to directory in data directory.

When reading files from directory, the external ID to reference will be an hash of
the full path. So chaning data type of "external ID" from numeric to string (model
field is already string).