Refactor Capabilities.app_revision() to support VERSION file
A VERSION file is generated for docker based setups to extract version
information. This was implemented by checking the revision log and parse
it.

The VERSION file contains the version (tag, commit counts, hash) and the
last commit log line. This line is parsed now in the same way.

Capabilities.app_revision() returns the version information depending
on what is available (VERSION file or git log).